avformat/avio: privatize source of truth for AVIOContext::written
Looking at 58190f57ac1a27037fb7aef76604476219f53fcc, the field
was supposed to be private, but during merging the field and the
group that had the comment about it got separated.

Thus, move the actual privately utilized state of this variable
into the private FFIOContext. Additionally, name the private field
somewhat better, so that it does not get confused with the amount
of bytes written out.